+++ src/app/product/models/product.model.ts
@@ -4,7 +4,8 @@ export class ProductModel {
         public name: string,
         public type: ProductType,
         public weight: number,
-        public cost: number
+        public cost: number,
+        public count: number
     ) {}
 }